requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
Werkzeug
Faker
